feat: expr matches woven elements
We've made the desicion to include woven elements in the `expr` since in most cases you don't care if an element was woven or not when selecting elements.

If you do care (and your widget extends `mu-jquery-widget/widget`) you can combine it with the close relative `mu-jquery-widget/expr` to mimic the old behaviour and find elements that have loom attributes but are not yet woven like this: ```$(":loom:not(:widget)"));```.

BREAKING CHANGE: `mu-jquery-loom/expr` now includes woven widgets.